LayoutTests/inspector/css/createStyleSheet-expected.txt

Test CSS.createStyleSheet.


== Running test suite: CSS.createStyleSheet
-- Running test case: CheckNoStyleSheets
PASS: Should be no stylesheets.

-- Running test case: CreateInspectorStyleSheetCall1
PASS: Should increase the list of stylesheets.
PASS: Added StyleSheet origin should be 'inspector'.
PASS: StyleSheet.isInspectorStyleSheet() should be true.
PASS: Added StyleSheet frame should be the main frame.

-- Running test case: CreateInspectorStyleSheetCall2
PASS: Should increase the list of stylesheets.
PASS: Added StyleSheet origin should be 'inspector'.
PASS: StyleSheet.isInspectorStyleSheet() should be true.
PASS: Added StyleSheet frame should be the main frame.
PASS: Added StyleSheet should be different from the last added stylesheet.

-- Running test case: CreateInspectorStyleSheetCall3
PASS: Should increase the list of stylesheets.
PASS: Added StyleSheet origin should be 'inspector'.
PASS: StyleSheet.isInspectorStyleSheet() should be true.
PASS: Added StyleSheet frame should be the main frame.
PASS: Added StyleSheet should be different from the last added stylesheet.
